requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
flake8==7.3.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
uritemplate==4.2.0
